chore(3dtiles): move b3dm mesh init from 3dTilesProvider to B3dmParser
